ISSUE-3 Separate scenario from login logic
Scenario is required to be in test to be detected.
